# docs/data/api.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Collections API
  description: Collections API
paths:
...
  /api/users/reset-password:
    post:
      summary: Reset password
      description: Reset password.
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/resetPasswordRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'success'
  /api/users/forgot-password:
    post:
      summary: Forgot password
      description: Forgot password.
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/forgotPasswordRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'success'

  ## #####################################
  ## Project Settings
  ## #####################################
...
  /api/models:
    get:
      summary: Get models
      description: Return models.
      operationId: getModels
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      $ref: '#/components/schemas/modelResponse'
      security:
        - bearerAuth: []
        - apiKeyAuth: []
  /api/models/{id}:
    get:
      summary: Get model
      description: Return model.
      operationId: getModel
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  model:
                    $ref: '#/components/schemas/modelResponse'
      security:
        - bearerAuth: []
        - apiKeyAuth: []

  ## #####################################
  ## Relations
  ## #####################################
...
  /api/models/{modelId}/fields:
    get:
      summary: Get fields
      description: Returns fields defined in the model in addition to id and timestamp.
      operationId: getFields
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  fields:
                    type: array
                    items:
                      $ref: '#/components/schemas/fieldResponse'
      security:
        - bearerAuth: []
        - apiKeyAuth: []

  ## #####################################
  ## Contents
  ## #####################################
...
  /api/roles/{id}/permissions:
    get:
      summary: Get permissions
      description: Return permissions.
      operationId: getPermissions
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  permissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/permissionResponse'
      security:
        - bearerAuth: []
        - apiKeyAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: access_token
  parameters:
    id:
      name: id
      in: path
      description: Primary Key
      required: true
      schema:
        $ref: '#/components/schemas/id'
...
  schemas:
...
    id:
      description: Primary Key
      example: 1
      type: integer
    createdAt:
      description: Created At
      example: '2023-09-29T00:53:20.604Z'
      type: string
    updatedAt:
      description: Updated At
      example: '2023-09-29T00:53:20.604Z'
      type: string
    token:
      description: Access Token
      example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
      type: string
...
    email:
      type: string
      format: email
      example: 'example@collections.dev'
    password:
      type: string
      format: password
      example: 'password'
...
    model:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/id'
        model:
          type: string
          example: 'posts'
        singleton:
          type: boolean
          example: false
        hidden:
          type: boolean
          example: false
        statusField:
          type: string
          example: 'status'
        draftValue:
          type: string
          example: 'draft'
        publishValue:
          type: string
          example: 'published'
        archiveValue:
          type: string
          example: 'archived'
        source:
          type: string
          example: null
        createdAt:
          $ref: '#/components/schemas/createdAt'
        updatedAt:
          $ref: '#/components/schemas/updatedAt'
    fieldOption:
      type: object
      properties:
        defaultValue:
          type: boolean
          example: true

    ## #####################################
    ## Authentications
    ## #####################################
...
    resetPasswordRequest:
      type: object
      required:
        - token
        - password
      properties:
        token:
          $ref: '#/components/schemas/token'
        password:
          $ref: '#/components/schemas/password'
    forgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          $ref: '#/components/schemas/email'

    ## #####################################
    ## Project Settings
    ## #####################################
...
    modelResponse:
      $ref: '#/components/schemas/model'

    ## #####################################
    ## Relations
    ## #####################################
...
    fieldResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/id'
        model:
          type: string
          example: 'zoo'
        modelId:
          $ref: '#/components/schemas/id'
        field:
          type: string
          example: 'name'
        label:
          type: string
          example: 'Name'
        special:
          type: string
          example: null
        interface:
          type: string
          example: 'input'
        options:
          type: string
          example: null
        readonly:
          type: boolean
          example: false
        required:
          type: boolean
          example: true
        hidden:
          type: boolean
          example: false
        sort:
          type: integer
          example: 1
        fieldOption:
          $ref: '#/components/schemas/fieldOption'
        createdAt:
          $ref: '#/components/schemas/createdAt'
        updatedAt:
          $ref: '#/components/schemas/updatedAt'

    ## #####################################
    ## Contents
    ## #####################################
...
    permissionResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/id'
        model:
          type: string
          example: 'zoo'
        modelId:
          $ref: '#/components/schemas/id'
        action:
          type: string
          example: 'read'
        roleId:
          $ref: '#/components/schemas/id'
        createdAt:
          $ref: '#/components/schemas/createdAt'
        updatedAt:
          $ref: '#/components/schemas/updatedAt'